+++ R/anovaWrapper.R
@@ -34,7 +34,7 @@
 #' @param barPlotSeparatePlots, By placing an independent variable in this box, different plots corresponding to the different levels of the independent variable will be displayed.
 #' @param contrastCi, By selecting this option, confidence intervals for the estimated mean difference and effect size will be included. By default the confidence level is set to 95%. This can be changed into the desired percentage.
 #'    Defaults to \code{FALSE}.
-#' @param contrastEffectSize, Include standardized mean differences, based on the effectsize function in the emmeans package.
+#' @param contrastEffectSize, Include standardized mean differences (Cohen's d). Contrast weights are rescaled so that their absolute values sum to 2, making the effect size invariant to the scale of the specified weights.
 #'    Defaults to \code{FALSE}.
 #' @param dependent, The variable of interest. This is also called the outcome variable.
 #' @param descriptivePlotErrorBar, Display error bars in the plot. The error bars can represent either confidence intervals or standard errors.